fix:修复编译报错

+++ app/port/luat_gpio_air101.c
@@ -101,7 +101,7 @@ int luat_gpio_setup(luat_gpio_t *gpio){
     return 0;
 }
 
-int luat_gpio_set(int pin, int level)
+int luat_gpio_set(int pin, uint8_t level)
 {
     if (pin < 0 || pin > WM_IO_PB_31) return 0;
     tls_gpio_write(pin, level);
